+++ R/analysis.R
@@ -276,34 +276,59 @@ oe_ht <- function(hic_matrix,
 
 #' Calculate compartment scores using Juicer tools
 #'
-#' @param hic_matrix Either a `hic_matrix` object, or the path of the `.hic`
-#'   name. If a file path, `resol` cannot be NULL since we can't infer
-#'   resolution directly from the `.hic` file.
+#' @param hic_matrix Either a `ht_table` object, or the path of the `.hic` name.
+#'   If a file path, `resol` cannot be NULL since we can't infer resolution
+#'   directly from the `.hic` file.
 #' @param chrom A character vector indicating which chromosomes to run. If
 #'   `NULL`, will calculate compartment scores for all chromosomes. However, in
 #'   this case, `hic_matrx` cannot be the `.hic` file name.
+#' @param resol Resolution of the matrix. Must be a positive integer when the
+#'   input is a Hi-C file.
+#' @param standard A standard compartment track which is used to "orient"
+#'   principal components. If `NULL`, no orienting will be performed. Can be the
+#'   name of a pre-computed reference track, e.g. gc.GRCh38.500kbp (see
+#'   `data(package = "hictools")`), or path to the bed file for the reference
+#'   track, or a `GRanges` object.
+#' @param oe The method to compute observed/expected values. `"juicer"`
+#'   indicates using using the values computed by Juicer tools. `"ht"` indicates
+#'   using hictools for the calculation. `"as-is"` indicates observed values
+#'   will be used as observed/expected ones, thus skipping the OE calculation.
+#'   This may be useful when the input is cofrag matrix rather than standard
+#'   Hi-C data. Won't take effect if `method` is `"juicer"`.
+#' @param norm The norm used for calculating compartment scores. If `hic_matrix`
+#'   is not a `.hic` file name, `norm` will be ignored.
 #' @export
 get_compartment <- function(hic_matrix,
                             method = c("juicer", "lieberman", "obs_exp", "nonzero", "average", "fanc"),
                             chrom = NULL,
                             standard = NULL,
-                            smooth = NULL,
-                            # juicertools = get_juicer_tools(),
-                            # java = "java",
+                            oe = c("juicer", "ht", "as-is"),
+                            juicertools = get_juicer_tools(),
+                            java = "java",
                             norm = c("NONE", "VC", "VC_SQRT", "KR", "SCALE"),
                             ...) {
   UseMethod("get_compartment")
 }
 
+
+#' Calculate compartment scores
+#' 
+#' Calculate compartment scores for Hi-C matrix in ht_table format.
+#' 
+#' @param hic_matrix A `ht_table` object for the Hi-C matrix.
+#' 
 #' @export
 get_compartment.ht_table <- function(hic_matrix,
                                      method = c("juicer", "lieberman", "obs_exp", "nonzero", "average", "fanc"),
                                      chrom = NULL,
                                      standard = NULL,
-                                     smooth = NULL,
+                                     oe = c("juicer", "ht", "as-is"),
+                                     juicertools = get_juicer_tools(),
+                                     java = "java",
                                      ...) {
   method <- match.arg(method)
-  
+  oe <- match.arg(oe)
+    
   if (is_null(chrom))
     chrom <- unique(c(hic_matrix$chrom1, hic_matrix$chrom2))
   else
@@ -311,37 +336,24 @@ get_compartment.ht_table <- function(hic_matrix,
   assert_that(is_character(chrom) && length(chrom) >= 1)
   
   resol <- attr(hic_matrix, "resol")
-  assert_that(is_scalar_integer(resol) &&
-                resol %in% c(500e3L, 1000e3L, 2500e3L, 5000e3L))
+  assert_that(is_valid_resol(resol) && resol >= 100e3L)
   
-  assert_that(is_null(standard) ||
-                is(standard, "GRanges") ||
-                is(standard, "data.frame"))
-  smooth <-
-    if (is_null(smooth))
-      NULL
-  else
-    as.integer(smooth)
-  assert_that(is_null(smooth) || smooth >= 1)
+  genome <- attr(hic, "genome")
   
-  genome <- attr(hic_matrix, "genome")
-  assert_that(is_null(genome) || is_scalar_character(genome))
-  
-  if (method %in% c("juicer", "fanc")) {
+  if (method == "juicer" || oe == "juicer") {
+    # Need to conver the input to .hic files if we need to calculate Juicer
+    # eigenvectors, or use Juicer tools to get oe scores.
+    
     # Save hic to a temporary .hic file, and call juicer tools to get compartments
     hic_file <- tempfile(fileext = ".hic")
     on.exit(file.remove(hic_file), add = TRUE)
     
-    args <- list(...)
-    java <- args$java %||% "java"
-    juicertools <- args$juicertools %||% get_juicer_tools()
-    
-    if (genome %in% c("hs37-1kg", "hg19", "GRCh37"))
+    if (genome %in% c("hs37-1kg", "hs37d5", "hg19") || grepl(pattern = "^GRCh37(\\.p[0-9]+)?$", x = genome))
       juicer_genome <- "hg19"
-    else if (genome %in% c("hg38", "GRCh38"))
+    else if (genome == "hg38" || grepl(pattern = "^GRCh38(\\.p[0-9]+)?$", x = genome))
       juicer_genome <- "hg38"
     else
-      stop(genome)
+      stop(pastep("Unknown genome: ", genome))
     
     hic_matrix %>% write_hic(
       file_path = hic_file,
@@ -351,57 +363,92 @@ get_compartment.ht_table <- function(hic_matrix,
       ref_genome = juicer_genome
     )
     
-    comps <- get_compartment.character(
-      hic_matrix = hic_file,
-      method = method,
-      chrom = chrom,
-      standard = standard,
-      smooth = 1L,
-      resol = attr(hic_matrix, "resol") %||% guess_resol(hic_matrix),
-      genome = attr(hic_matrix, "genome") %||% args$genome,
-      juicertools = juicertools,
-      java = java,
-      norm = args$norm %||% "NONE"
-    )
+    if (method == "juicer")
+      comps <- get_compartment.character(
+        hic_matrix = hic_file,
+        method = method,
+        chrom = chrom,
+        standard = standard,
+        resol = attr(hic_matrix, "resol") %||% guess_resol(hic_matrix),
+        genome = attr(hic_matrix, "genome") %||% args$genome,
+        juicertools = juicertools,
+        java = java,
+        norm = "NONE"
+      )
+    else {
+      comps <- chrom %>%
+        map(function(chrom) {
+            load_juicer_hic(
+              file_path = hic_file,
+              chrom = chrom,
+              resol = resol,
+              type = "oe",
+              norm = "NONE",
+              genome = genome
+            )
+        }) %>% 
+        concat_hic() %>%
+        get_compartment.ht_table(
+          method = method,
+          chrom = chrom,
+          standard = standard,
+          oe = "as-is",
+          juicertools = juicertools,
+          java = java,
+          ...
+        )
+    }
   } else {
     comps <- compartment_ht(
       hic_matrix = hic_matrix,
       method = method,
       chrom = chrom,
       npc = 2L,
-      standard = standard
+      oe = oe
     )
   }
   
-  if (isTRUE(smooth > 1))
-    comps$score <- zoo::rollmean(
-      comps$score,
-      k = smooth,
-      na.pad = TRUE,
-      na.rm = TRUE,
-      align = "center"
-    )
-  
+  if (!is_null(standard)) {
+    # PC1/2/3, ..., or simply score
+    score_cols <- colnames(mcols(comps)) %>% grep(pattern = "^PC[0-9]+", value = TRUE)
+    if (length(score_cols) == 0)
+      score_cols <- colnames(mcols(comps))[1]
+    
+    comps <- orient_compartment(comps, standard, score_cols = score_cols)
+  }
+
   return(comps)
 }
 
 
+#' Calculate compartment scores
+#' 
+#' Load Hi-C matrix from .hic file and calculate compartment scores.
+#' 
+#' @param genome Name of the reference genome.
+#' @param resol A positive integer for resolution.
 #' @export
 get_compartment.character <- function(hic_matrix,
                                       method = c("juicer", "lieberman", "obs_exp", "nonzero", "average", "fanc"),
                                       chrom = NULL,
-                                      standard = NULL,
-                                      smooth = NULL,
                                       resol,
+                                      standard = NULL,
+                                      oe = c("juicer", "ht", "as-is"),
                                       genome = NULL,
                                       juicertools = get_juicer_tools(),
                                       java = "java",
-                                      norm = c("NONE", "VC", "VC_SQRT", "KR", "SCALE")) {
-  
+                                      norm = c("NONE", "VC", "VC_SQRT", "KR", "SCALE"),
+                                      ...) {
+  assert_that(grepl("\\.hic$", x = hic_matrix) && assertthat::is.readable(hic_matrix))
   method <- match.arg(method)
+  oe <- match.arg(oe)
   norm <- match.arg(norm)
   
-  if (isTRUE(method == "juicer")) {
+  assert_that(is_character(chrom) && length(chrom) >= 1)
+  assert_that(is_valid_resol(resol) && resol >= 100e3L)
+  assert_that(!is_null(bedtorch::get_seqinfo(genome = genome)))
+  
+  if (method == "juicer") {
     # No need to load the .hic file. Instead, directly call compartments using Juicer tools
     comps <- compartment_juicer_file(
       hic_file = hic_matrix,
@@ -410,70 +457,102 @@ get_compartment.character <- function(hic_matrix,
       norm = norm,
       chrom = chrom,
       resol = resol,
-      standard = standard,
       genome = genome
     )
   } else if (method == "fanc") {
-    # No need to load the .hic file. Instead, directly call compartments using Juicer tools
-    comps <- compartment_fanc(
-      hic_file = hic_matrix,
-      norm = norm,
-      resol = resol,
-      genome = genome
-    )
+    stop("FAN-C compartment is not supported at this point")
+    # # No need to load the .hic file. Instead, directly call compartments using Juicer tools
+    # comps <- compartment_fanc(
+    #   hic_file = hic_matrix,
+    #   norm = norm,
+    #   resol = resol,
+    #   genome = genome
+    # )
   } else {
-    hic_matrix <-
-      load_juicer_hic(
-        file_path = hic_matrix,
-        resol = resol,
+    if (oe == "juicer") {
+      juicer_type <- "oe"
+      oe <- "as-is"
+    } else {
+      juicer_type <- "observed"
+    }
+    
+    browser()
+    comps <- chrom %>%
+      map(function(chrom) {
+        load_juicer_hic(
+          file_path = hic_matrix,
+          chrom = chrom,
+          resol = resol,
+          # Which type to load?
+          type = juicer_type, 
+          norm = "NONE",
+          genome = genome
+        )
+      }) %>% 
+      concat_hic() %>%
+      get_compartment.ht_table(
+        method = method,
         chrom = chrom,
-        type = "observed",
-        norm = norm, 
-        genome = genome
+        standard = standard,
+        oe = oe,
+        juicertools = juicertools,
+        java = java,
+        ...
       )
-    comps <- get_compartment(
-      hic_matrix = hic_matrix,
-      method = method,
-      chrom = chrom,
-      standard = standard,
-      smooth = 1L
-    )
+    
+    
+    # 
+    # 
+    # hic_matrix <-
+    #   load_juicer_hic(
+    #     file_path = hic_matrix,
+    #     resol = resol,
+    #     chrom = chrom,
+    #     # Which type to load?
+    #     type = switch(oe, juicer = "oe", "observed"),
+    #     norm = norm, 
+    #     genome = genome
+    #   )
+    # comps <- get_compartment(
+    #   hic_matrix = hic_matrix,
+    #   method = method,
+    #   chrom = chrom,
+    #   standard = standard,
+    #   oe = switch(oe, juicer = "as-is", oe)
+    # )
   }
   
-  if (isTRUE(smooth > 1)) {
-    comps$score <- zoo::rollmean(
-      comps$score,
-      k = smooth,
-      na.pad = TRUE,
-      na.rm = TRUE,
-      align = "center"
-    )
-  }
   
+  if (!is_null(standard)) {
+    # PC1/2/3, ..., or simply score
+    score_cols <- colnames(mcols(comps)) %>% grep(pattern = "^PC[0-9]+", value = TRUE)
+    if (length(score_cols) == 0)
+      score_cols <- colnames(mcols(comps))[1]
+    
+    comps <- orient_compartment(comps, standard, score_cols = score_cols)
+  }
   
   return(comps)
 }
 
 
-#' Get compartment using Juicer tools, from an existing .hic file
+#' Compute compartment scores using Juicer tools, from an existing .hic file
 compartment_juicer_file <- function(hic_file,
                                     juicertools = get_juicer_tools(),
                                     java = "java",
                                     norm = c("NONE", "VC", "VC_SQRT", "KR", "SCALE"),
                                     chrom = NULL,
                                     resol = NULL,
-                                    standard = NULL,
                                     genome = NULL) {
-  assert_that(is_scalar_character(hic_file) && endsWith(hic_file, ".hic"))
+  assert_that(grepl(pattern = "\\.hic$", x = hic_file) && assertthat::is.readable(hic_file))
   assert_that(is_scalar_character(java))
   assert_that(is_scalar_character(juicertools))
   norm <- match.arg(norm)
   chrom <- as.character(chrom)
   assert_that(length(chrom) >= 1)
   resol <- as.integer(resol)
-  assert_that(is_scalar_integer(resol) && resol %in% c(500e3L, 1000e3L, 2500e3L, 5000e3L))
-  assert_that(is_null(standard) || is(standard, "GRanges") || is(standard, "data.frame"))
-  assert_that(is_null(genome) || is_scalar_character(genome))
+  assert_that(is_valid_resol(resol) && resol >= 100e3L)
+  assert_that(is_null(genome) || (is_scalar_character(genome) && !is_null(bedtorch::get_seqinfo(genome))))
   
   comps <- chrom %>% map(function(chrom) {
     ev_file <- tempfile()
@@ -499,11 +578,114 @@ compartment_juicer_file <- function(hic_file,
   
   comps <- suppressWarnings(bedtorch::as.GenomicRanges(comps[!is.na(score)], genome = genome) %>% GenomicRanges::trim())
   
-  if (!is_null(standard))
-    comps %<>% normalize_compartment(standard = standard)
-  
   return(comps)
 }
+
+
+#' Compute compartment scores using hictools's own implementation of compartment analysis, as opposed to juicer
+#' 
+#' @param npc A positive integer indicating how many principal components to keep.
+compartment_ht <-
+  function(hic_matrix,
+           method = c("lieberman", "obs_exp", "nonzero", "average"),
+           oe = c("ht", "as-is"),
+           chrom = NULL,
+           npc = 2L) {
+    assert_that(is(hic_matrix, "ht_table"))
+    method <- match.arg(method)
+    oe <- match.arg(oe)
+    assert_that(is_scalar_integer(npc) && npc >= 1)
+    
+    if (is_null(chrom))
+      chrom <-
+      unique(c(
+        as.character(hic_matrix$chrom1),
+        as.character(hic_matrix$chrom2)
+      ))
+    assert_that(is_character(chrom) && length(chrom) >= 1)
+    
+    resol <- attr(hic_matrix, "resol")
+    assert_that(is_valid_resol(resol) && resol >= 100e3L)
+    type <- attr(hic_matrix, "type")
+    assert_that(is_scalar_character(type))
+    
+    genome <- attr(hic_matrix, "genome")
+    
+    if (oe != "as-is")
+      hic_matrix <- oe_ht(hic_matrix, method = method)
+    
+    comps <- chrom %>% map(function(chrom) {
+      hic_matrix <- hic_matrix[chrom1 == chrom & chrom2 == chrom]
+      pos_start <- min(c(hic_matrix$pos1, hic_matrix$pos2))
+      
+      mc <- hic_matrix %>%
+        convert_hic_matrix() %>%
+        cor(use = "pairwise.complete.obs")
+      
+      # We need to exclude rows/columns which are all NAs
+      all_na <- apply(mc, MARGIN = 1, FUN = function(x) all(is.na(x)))
+      
+      # Sometimes, a single data point can be NA. This is different from the above case,
+      # where the whole row/column is NA
+      # We need to set these data points to 0 for PCA to work
+      mc2 <- mc[!all_na, !all_na]
+      mc2[which(is.na(mc2))] <- 0
+      pc_subset <- prcomp(mc2, scale. = TRUE) %>% .$rotation %>% .[, 1:npc]
+      pc <- matrix(rep(NA, ncol(mc) * npc), ncol = npc)
+      pc[!all_na, ] <- pc_subset
+      
+      result <- data.table::data.table(
+        chrom = chrom,
+        start = as.integer((seq_along(pc[, 1]) - 1) * resol + pos_start)
+      )
+      result[, `:=`(end = as.integer(start + resol), score = pc[, 1])]
+      result[, score := ifelse(is.infinite(score) | is.nan(score), NA, score)]
+      data.table::setkey(result, "chrom", "start", "end")
+      
+      seq.int(npc) %>% walk(function(pc_idx) {
+        result[, (paste0("PC", pc_idx)) := pc[, pc_idx]]
+      })
+      
+      result
+    }) %>%
+      data.table::rbindlist()
+    
+    return(suppressWarnings(comps %>% 
+                                bedtorch::as.GenomicRanges(genome = genome) %>% 
+                                GenomicRanges::trim()))
+    
+    # gc_track <- suppressWarnings({
+    #   local({
+    #     data_env <- env()
+    #     data_name <- str_interp("gc.${genome}.${resol%/%1000}kbp")
+    #     data(list = data_name, package = "hictools", envir = data_env)
+    #     data_env[[data_name]]
+    #   })
+    # })
+    # gc_track$score <- gc_track$gc
+    # gc_track$gc <- NULL
+    # 
+    # comps <- suppressWarnings(comps %>% bedtorch::as.GenomicRanges() %>% GenomicRanges::trim())
+    # 
+    # if (!is_null(gc_track)) {
+    #   comps <- orient_compartment(comps, standard = gc_track, score_cols = colnames(mcols(comps)))
+    #   
+    #   score <- comps$score
+    #   comps$score <- NULL
+    #   mcols(comps) <- cbind(data.frame(score = score), mcols(comps))
+    # } else {
+    #   score <- mcols(comps)[, 1]
+    #   comps$score <- NULL
+    #   mcols(comps) <- cbind(data.frame(score = score), mcols(comps))
+    # }
+    # 
+    # 
+    # if (!is_null(standard))
+    #   comps %<>% orient_compartment(standard = standard, score_cols = paste0("PC", seq.int(npc)))
+    # # comps %<>% orient_compartment(standard = standard)
+    
+    # suppressWarnings(comps %>% bedtorch::as.GenomicRanges() %>% GenomicRanges::trim())
+  }
 
 
 #' Get compartment using FANC, from an existing .hic file
@@ -515,7 +697,7 @@ compartment_fanc <- function(hic_file,
   assert_that(is_scalar_character(hic_file) && endsWith(hic_file, ".hic"))
   norm <- match.arg(norm)
   resol <- as.integer(resol)
-  assert_that(is_scalar_integer(resol) && resol %in% c(250e3L, 500e3L, 1000e3L))
+  assert_that(is_valid_resol(resol) && resol >= 100e3L)
   assert_that(is_null(genome) || is_scalar_character(genome))
   
   ab_file <- tempfile(fileext = ".ab")
@@ -574,7 +756,7 @@ compartment_fanc <- function(hic_file,
   gc_track$gc <- NULL
   
   if (!is_null(gc_track)) {
-    comps <- normalize_compartment(comps, standard = gc_track, score_cols = colnames(mcols(comps)))
+    comps <- orient_compartment(comps, standard = gc_track, score_cols = colnames(mcols(comps)))
 
     score <- comps$score
     comps$score <- NULL
@@ -651,102 +833,96 @@ pearson_ht <- function(hic_matrix, chrom, method = "lieberman") {
 }
 
 
-#' Flip the compartment score signs according to `standard`
+#' Orient the compartment score signs according to `standard`
 #'
-#' @param compartment A `GRanges` or `data.frame` containing compartment scores.
-#' @param standard A `GRanges` or `data.frame` of standard compartment scores.
+#' @param compartment A `GRanges` object for compartment scores.
+#' @param standard A `GRanges` for the standard compartment track.
+#' @param score_cols A character vector indicating which columns are candidates
+#'   for compartment scores. For example, `c("PC1", "PC2")` means from `PC1` and
+#'   `PC2`, picking the one most likely to be associated with A/B compartments,
+#'   and orient it if necessary.
 #' @return The compartment. The signs of compartment scores may have been
 #'   flipped based on `standard`.
 #' @export
-flip_compartment <- function(compartment, standard) {
-  UseMethod("flip_compartment")
-}
-
-#' @export
-flip_compartment.data.frame <- function(compartment, standard) {
-  suppressWarnings(compartment %<>% bedtorch::as.GenomicRanges() %>% GenomicRanges::trim())
-  compartment <- flip_compartment(compartment, standard)
-  bedtorch::as.bedtorch_table(compartment)
-}
-
-#' @export
-flip_compartment.GRanges <- function(compartment, standard) {
-  if (inherits(standard, "data.frame"))
-    suppressWarnings(standard %<>% bedtorch::as.GenomicRanges() %>% GenomicRanges::trim())
+orient_compartment <- function(compartment, standard, score_cols = c("score")) {
+  assert_that(is_character(score_cols) && length(score_cols) >= 1)
+  assert_that(is(compartment, "GRanges"))
+  assert_that(is(standard, "GRanges"))
   
-  compartment <- unique(GenomicRanges::seqnames(compartment)) %>%
-    map(function(chrom) {
-      compartment <- compartment[GenomicRanges::seqnames(compartment) == chrom]
-      hits <- GenomicRanges::findOverlaps(compartment, standard)
-      
-      intervals_1 <- compartment[S4Vectors::queryHits(hits)]
-      intervals_2 <- standard[S4Vectors::subjectHits(hits)]
-      
-      correlation <- cor(intervals_1$score, intervals_2$score, use = "complete.obs")
-      if (correlation < 0)
-        compartment$score <- compartment$score * (-1)
-      
-      compartment
-    })
+  # Check for compatibility
+  resol1 <- local({
+    dt <- compartment %>% bedtorch::as.bedtorch_table()
+    dt[, dplyr::lead(start) - start, by = "chrom"][, min(V1, na.rm = TRUE)]
+  })
+  resol2 <- local({
+    dt <- standard %>% bedtorch::as.bedtorch_table()
+    dt[, dplyr::lead(start) - start, by = "chrom"][, min(V1, na.rm = TRUE)]
+  })
+  assert_that(is_valid_resol(resol1) && is_valid_resol(resol2) && resol1 == resol2)
   
-  rlang::exec(c, !!!compartment)
-}
-
-
-#' Normalize the compartment score signs according to `standard`
-#'
-#' @param compartment A `GRanges` or `data.frame` containing compartment scores.
-#' @param standard A `GRanges` or `data.frame` of standard compartment scores.
-#' @return The compartment. The signs of compartment scores may have been
-#'   flipped based on `standard`.
-#' @export
-normalize_compartment <- function(compartment, standard, score_cols = c("score")) {
-  UseMethod("normalize_compartment")
-}
-
-#' @export
-normalize_compartment.data.frame <- function(compartment, standard, score_cols = c("score")) {
-  suppressWarnings(compartment %<>% bedtorch::as.GenomicRanges() %>% GenomicRanges::trim())
-  compartment <- normalize_compartment(compartment, standard, score_cols = score_cols)
-  bedtorch::as.bedtorch_table(compartment)
-}
-
-#' @export
-normalize_compartment.GRanges <- function(compartment, standard, score_cols = c("score")) {
-  if (inherits(standard, "data.frame"))
-    suppressWarnings(standard %<>% bedtorch::as.GenomicRanges() %>% GenomicRanges::trim())
-
-  compartment <- unique(GenomicRanges::seqnames(compartment)) %>%
+  # Make compatible genomes
+  original_seqinfo <- seqinfo(compartment)
+  seqlevels(compartment) <- seqlevels(standard)
+  seqinfo(compartment) <- seqinfo(standard)
+  
+  # By default, the first column of standard is used for determining the orientation
+  standard$score <- mcols(standard)[[1]]
+  
+  compartment <- unique(seqnames(compartment)) %>%
     map(function(chrom) {
-      compartment <- compartment[GenomicRanges::seqnames(compartment) == chrom]
+      compartment <- compartment[seqnames(compartment) == chrom]
       
       # Calcualte the correlation with PC1, PC2, etc...
       cor_values <- map_dbl(score_cols, function(score_col) {
         compartment$score <- mcols(compartment)[[score_col]]
-        hits <- GenomicRanges::findOverlaps(compartment, standard)
-        intervals_1 <- compartment[S4Vectors::queryHits(hits)]
-        intervals_2 <- standard[S4Vectors::subjectHits(hits)]
+        hits <- findOverlaps(compartment, standard)
+        intervals_1 <- compartment[queryHits(hits)]
+        intervals_2 <- standard[subjectHits(hits)]
         cor(intervals_1$score, intervals_2$score, use = "complete.obs")
       })
+      names(cor_values) <- score_cols
       
-      # Pick the one with the highest correlation coefficient, regardless of the sign
-      # PC2 should be sufficiently larger than PC1 to be picked as compartment scores
-      pick_index <- which.max(abs(cor_values))
-      if (isTRUE(pick_index != 1) && abs(cor_values[pick_index]) - abs(cor_values[1]) > 0.5) {
-        correlation_str <- paste(abs(cor_values), collapse = ", ")
-        logging::logwarn(str_interp("Compartment for chromosome ${chrom}: choosing PC${pick_index}, correlation: ${correlation_str}"))
-      } else
-        pick_index <- 1
+      # Only pick higher-order PCs when the correlation is sufficiently high (> 0.5) and
+      # sufficiently larger than PC1 correlation (delta > 0.1)
+      min_higher_pc_cor <- 0.5
+      min_delta_pc_cor <- 0.1
       
-      if (cor_values[pick_index] < 0)
-        compartment$score <- mcols(compartment)[[score_cols[pick_index]]] * (-1)
-      else
-        compartment$score <- mcols(compartment)[[score_cols[pick_index]]]
+      sorted_cor <- sort(abs(cor_values), decreasing = TRUE)
+      
+      if (names(sorted_cor[1]) != score_cols[1] &&
+          sorted_cor[1] > min_higher_pc_cor &&
+          sorted_cor[1] - sorted_cor[score_cols[1]] > min_delta_pc_cor) {
+        # Pick the one with the highest correlation coefficient, regardless of the sign
+        # PC2 should be sufficiently larger than PC1 to be picked as compartment scores
+        
+        pick_cor_name <- names(sorted_cor[1])
+        flip_ev <- cor_values[pick_cor_name] < 0
+        logging::logwarn(
+          str_interp(
+            "Compartment for chromosome ${chrom}: higher PC ${pick_cor_name} is chosen. Correlation: ${cor_values[pick_cor_name]}"
+          )
+        )
+      } else {
+        pick_cor_name <- score_cols[1]
+        flip_ev <- cor_values[pick_cor_name] < 0
+      }
+      
+      mcols_comps <- mcols(compartment)
+      score <- mcols_comps[[pick_cor_name]]
+      if (flip_ev)
+        score <- -score
+      
+      mcols_comps$score <- NULL
+      mcols_comps <- cbind(list(score = score), mcols_comps)
+      mcols(compartment) <- mcols_comps
       
       return(compartment)
     })
   
-  rlang::exec(c, !!!compartment)
+  compartment <- rlang::exec(c, !!!compartment)
+  seqlevels(compartment) <- seqlevels(original_seqinfo)
+  seqinfo(compartment) <- original_seqinfo
+  return(compartment)
 }
 
 
@@ -822,110 +998,6 @@ hicexplorer_pca <- function(hic_matrix,
     unlink(c(hic_cool, pc_bedgraph, pm, oem))
   })
 }
-
-
-# hictools's own implementation of compartment analysis, as opposed to juicer
-compartment_ht <-
-  function(hic_matrix,
-           method = c("lieberman", "obs_exp", "nonzero", "average"),
-           chrom = NULL,
-           npc = 2L,
-           genome = c("hs37-1kg", "GRCh37", "GRCh38")) {
-    assert_that(is(hic_matrix, "ht_table"))
-    method <- match.arg(method)
-    assert_that(is_scalar_integer(npc) && npc >= 1)
-    
-    assert_that(is_null(standard) ||
-                  is(standard, "GRanges") || is(standard, "data.frame"))
-    
-    if (is_null(chrom))
-      chrom <-
-        unique(c(
-          as.character(hic_matrix$chrom1),
-          as.character(hic_matrix$chrom2)
-        ))
-    assert_that(is_character(chrom) && length(chrom) >= 1)
-    
-    resol <- attr(hic_matrix, "resol")
-    assert_that(is_scalar_integer(resol) && resol > 0)
-    type <- attr(hic_matrix, "type")
-    assert_that(is_scalar_character(type))
-    
-    if (isTRUE(type == "observed"))
-      hic_matrix <- oe_ht(hic_matrix, method = method)
-    
-    comps <- chrom %>% map(function(chrom) {
-      hic_matrix <- hic_matrix[chrom1 == chrom & chrom2 == chrom]
-      pos_start <- min(c(hic_matrix$pos1, hic_matrix$pos2))
-      
-      mc <- hic_matrix %>%
-        convert_hic_matrix() %>%
-        cor(use = "pairwise.complete.obs")
-      
-      # We need to exclude rows/columns which are all NAs
-      all_na <-
-        seq.int(ncol(mc)) %>% map_lgl(function(idx) {
-          all(is.na(mc[, idx]))
-        })
-      
-      # Sometimes, a single data point can be NA. This is different from the above case,
-      # where the whole row/column is NA
-      # We need to set these data points to 0 for PCA to work
-      mc2 <- mc[!all_na,!all_na]
-      mc2[which(is.na(mc2))] <- 0
-      pc_subset <- prcomp(mc2, scale. = TRUE) %>% .$rotation %>% .[, 1:npc]
-      pc <- matrix(rep(NA, ncol(mc) * npc), ncol = npc)
-      pc[!all_na, ] <- pc_subset
-      
-      result <- data.table::data.table(
-        chrom = chrom,
-        start = as.integer((seq_along(pc[, 1]) - 1) * resol + pos_start)
-      )
-      result[, `:=`(end = as.integer(start + resol), score = pc[, 1])]
-      result[, score := ifelse(is.infinite(score) | is.nan(score), NA, score)]
-      data.table::setkey(result, "chrom", "start", "end")
-      
-      seq.int(npc) %>% walk(function(pc_idx) {
-        result[, (paste0("PC", pc_idx)) := pc[, pc_idx]]
-      })
-      
-      result
-    }) %>%
-      data.table::rbindlist() %>%
-      bedtorch::as.bedtorch_table(attr(hic_matrix, "genome"))
-    
-    gc_track <- suppressWarnings({
-      local({
-        data_env <- env()
-        data_name <- str_interp("gc.${genome}.${resol%/%1000}kbp")
-        data(list = data_name, package = "hictools", envir = data_env)
-        data_env[[data_name]]
-      })
-    })
-    gc_track$score <- gc_track$gc
-    gc_track$gc <- NULL
-    
-    comps <- suppressWarnings(comps %>% bedtorch::as.GenomicRanges() %>% GenomicRanges::trim())
-    
-    if (!is_null(gc_track)) {
-      comps <- normalize_compartment(comps, standard = gc_track, score_cols = colnames(mcols(comps)))
-      
-      score <- comps$score
-      comps$score <- NULL
-      mcols(comps) <- cbind(data.frame(score = score), mcols(comps))
-    } else {
-      score <- mcols(comps)[, 1]
-      comps$score <- NULL
-      mcols(comps) <- cbind(data.frame(score = score), mcols(comps))
-    }
-    
-    
-    if (!is_null(standard))
-      comps %<>% normalize_compartment(standard = standard, score_cols = paste0("PC", seq.int(npc)))
-      # comps %<>% normalize_compartment(standard = standard)
-      
-    suppressWarnings(comps %>% bedtorch::as.GenomicRanges() %>% GenomicRanges::trim())
-  }
 
 
 #' Calculate correlation coefficients between two compartment tracks