+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2021-11-14
+
+Version: v0.1.5.1
+
+Changes:
+
+* Improve compartment analysis
+* Improve compartment orientation flipping
+
 ## 2021-11-08
 
 Version: v0.1.4.3